# apps/api/docs/swagger.yaml
definitions:
...
  dto.AIUsageSnapshot:
    properties:
      token_usage:
        type: integer
      total_requests:
        type: integer
    type: object
...
  dto.AdminUser:
    properties:
      email:
        type: string
      id:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
  dto.AnalyticsSnapshot:
    properties:
      active_users:
        type: integer
      total_activities:
        type: integer
      total_courses:
        type: integer
      total_users:
        type: integer
    type: object
  dto.BillingHistoryItem:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      plan_name:
        type: string
      status:
        type: string
    type: object
  dto.BillingPlan:
    properties:
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  dto.Course:
    properties:
      description:
        type: string
      domain:
        type: string
      id:
        type: string
      level:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  dto.SubscribeRequest:
    properties:
      plan_id:
        type: string
    required:
    - plan_id
    type: object
...
  dto.UpdateUserStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
  dto.UpsertCourseRequest:
    properties:
      description:
        type: string
      domain:
        type: string
      level:
        type: string
      status:
        type: string
      title:
        type: string
    required:
    - domain
    - title
    type: object
...
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      request_id:
        type: string
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: integer
      fields:
        additionalProperties:
          type: string
        description: validation field errors
        type: object
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /admin/ai-usage:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.AIUsageSnapshot'
              type: object
      security:
      - BearerAuth: []
      summary: Get admin AI usage
      tags:
      - admin
  /admin/analytics:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.AnalyticsSnapshot'
              type: object
      security:
      - BearerAuth: []
      summary: Get admin analytics
      tags:
      - admin
  /admin/courses:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by title or description
        in: query
        name: q
        type: string
      - description: Filter by domain
        in: query
        name: domain
        type: string
      - description: Filter by level
        in: query
        name: level
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Course'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List admin courses
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Course payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.Course'
              type: object
      security:
      - BearerAuth: []
      summary: Create admin course
      tags:
      - admin
  /admin/courses/{id}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete admin course
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Course payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.Course'
              type: object
      security:
      - BearerAuth: []
      summary: Update admin course
      tags:
      - admin
  /admin/users:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by email or name
        in: query
        name: q
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AdminUser'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List admin users
      tags:
      - admin
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminUser'
              type: object
      security:
      - BearerAuth: []
      summary: Update user status
      tags:
      - admin
...
  /billing/history:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by plan name
        in: query
        name: q
        type: string
      - description: Filter by payment status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.BillingHistoryItem'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get billing history
      tags:
      - billing
  /billing/plans:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by plan name or description
        in: query
        name: q
        type: string
      - description: Filter by currency
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.BillingPlan'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List billing plans
      tags:
      - billing
  /billing/subscribe:
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SubscribeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Subscribe billing plan
      tags:
      - billing
...
swagger: "2.0"